requirements.txt
Flask==3.0.3
joblib==1.4.2
numpy==2.0.0
pandas==2.2.2
pyngrok==7.1.6
python-dotenv==1.0.1
scikit_learn==1.5.0
Requests==2.31.0
xgboost
waitress
ngrok